260718 커밋

+++ src/main/java/com/example/fund/fund/service/FundHoldingService.java
@@ -14,6 +14,6 @@ public class FundHoldingService {
     private final FundHoldingRepository fundHoldingRepository;
 
     public List<FundHolding> getHoldingsByUserId(int userId) {
-        return fundHoldingRepository.findByUser_UserIdOrderByJoinedAtDesc(userId);
+        return fundHoldingRepository.findByUser_UserIdOrderByJoinedAtDesc((long) userId);
     }
 }
